refactor: 뷰 파일을 web폴더로 분리
